pyproject.toml
[project]
name = "audio-to-mermaid"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "anthropic>=0.40.0",
    "pynput>=1.7.6",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
